entities and JPA controllers done. No tuve tiempo de revisar todos los constructores y así pero me aseguré de que las relaciones estén.

+++ src/java/model/OfferJpaController.java
@@ -4,9 +4,7 @@
  */
 package model;
 
-import model.exceptions.NonexistentEntityException;
-import model.exceptions.PreexistingEntityException;
-import model.exceptions.RollbackFailureException;
+import Datastore.Entities.Offer;
 import java.io.Serializable;
 import java.util.List;
 import javax.persistence.EntityManager;
@@ -15,14 +13,15 @@ import javax.persistence.Query;
 import javax.persistence.EntityNotFoundException;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Root;
+import Datastore.Entities.Vehicleadvert;
+import java.util.ArrayList;
+import java.util.Collection;
+import Datastore.Entities.Businessadvert;
 import javax.transaction.UserTransaction;
-import Datastore.Entities.Administrato;
-import Datastore.Entities.Offer;
-import javax.annotation.Resource;
-import javax.naming.Context;
-import javax.naming.InitialContext;
-import javax.persistence.Persistence;
-import javax.persistence.PersistenceUnit;
+import model.exceptions.IllegalOrphanException;
+import model.exceptions.NonexistentEntityException;
+import model.exceptions.PreexistingEntityException;
+import model.exceptions.RollbackFailureException;
 
 /**
  *
@@ -30,34 +29,58 @@ import javax.persistence.PersistenceUnit;
  */
 public class OfferJpaController implements Serializable {
 
-    public OfferJpaController() {
-
+    public OfferJpaController(UserTransaction utx, EntityManagerFactory emf) {
+        this.utx = utx;
+        this.emf = emf;
     }
-    
-    @Resource private UserTransaction utx;
-    @PersistenceUnit private EntityManagerFactory emf = Persistence.createEntityManagerFactory("ProjectPU");
+    private UserTransaction utx = null;
+    private EntityManagerFactory emf = null;
 
     public EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
 
     public void create(Offer offer) throws PreexistingEntityException, RollbackFailureException, Exception {
+        if (offer.getVehicleadvertCollection() == null) {
+            offer.setVehicleadvertCollection(new ArrayList<Vehicleadvert>());
+        }
+        if (offer.getBusinessadvertCollection() == null) {
+            offer.setBusinessadvertCollection(new ArrayList<Businessadvert>());
+        }
         EntityManager em = null;
-        Context ctx = new InitialContext();
-        this.utx = (UserTransaction) ctx.lookup("java:comp/UserTransaction");
-        
         try {
             utx.begin();
             em = getEntityManager();
-            Administrato adminID = offer.getAdminID();
-            if (adminID != null) {
-                adminID = em.getReference(adminID.getClass(), adminID.getAdminID());
-                offer.setAdminID(adminID);
+            Collection<Vehicleadvert> attachedVehicleadvertCollection = new ArrayList<Vehicleadvert>();
+            for (Vehicleadvert vehicleadvertCollectionVehicleadvertToAttach : offer.getVehicleadvertCollection()) {
+                vehicleadvertCollectionVehicleadvertToAttach = em.getReference(vehicleadvertCollectionVehicleadvertToAttach.getClass(), vehicleadvertCollectionVehicleadvertToAttach.getCode());
+                attachedVehicleadvertCollection.add(vehicleadvertCollectionVehicleadvertToAttach);
             }
+            offer.setVehicleadvertCollection(attachedVehicleadvertCollection);
+            Collection<Businessadvert> attachedBusinessadvertCollection = new ArrayList<Businessadvert>();
+            for (Businessadvert businessadvertCollectionBusinessadvertToAttach : offer.getBusinessadvertCollection()) {
+                businessadvertCollectionBusinessadvertToAttach = em.getReference(businessadvertCollectionBusinessadvertToAttach.getClass(), businessadvertCollectionBusinessadvertToAttach.getCode());
+                attachedBusinessadvertCollection.add(businessadvertCollectionBusinessadvertToAttach);
+            }
+            offer.setBusinessadvertCollection(attachedBusinessadvertCollection);
             em.persist(offer);
-            if (adminID != null) {
-                adminID.getOfferCollection().add(offer);
-                adminID = em.merge(adminID);
+            for (Vehicleadvert vehicleadvertCollectionVehicleadvert : offer.getVehicleadvertCollection()) {
+                Offer oldOfferOfVehicleadvertCollectionVehicleadvert = vehicleadvertCollectionVehicleadvert.getOffer();
+                vehicleadvertCollectionVehicleadvert.setOffer(offer);
+                vehicleadvertCollectionVehicleadvert = em.merge(vehicleadvertCollectionVehicleadvert);
+                if (oldOfferOfVehicleadvertCollectionVehicleadvert != null) {
+                    oldOfferOfVehicleadvertCollectionVehicleadvert.getVehicleadvertCollection().remove(vehicleadvertCollectionVehicleadvert);
+                    oldOfferOfVehicleadvertCollectionVehicleadvert = em.merge(oldOfferOfVehicleadvertCollectionVehicleadvert);
+                }
+            }
+            for (Businessadvert businessadvertCollectionBusinessadvert : offer.getBusinessadvertCollection()) {
+                Offer oldOfferOfBusinessadvertCollectionBusinessadvert = businessadvertCollectionBusinessadvert.getOffer();
+                businessadvertCollectionBusinessadvert.setOffer(offer);
+                businessadvertCollectionBusinessadvert = em.merge(businessadvertCollectionBusinessadvert);
+                if (oldOfferOfBusinessadvertCollectionBusinessadvert != null) {
+                    oldOfferOfBusinessadvertCollectionBusinessadvert.getBusinessadvertCollection().remove(businessadvertCollectionBusinessadvert);
+                    oldOfferOfBusinessadvertCollectionBusinessadvert = em.merge(oldOfferOfBusinessadvertCollectionBusinessadvert);
+                }
             }
             utx.commit();
         } catch (Exception ex) {
@@ -77,29 +100,72 @@ public class OfferJpaController implements Serializable {
         }
     }
 
-    public void edit(Offer offer) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void edit(Offer offer) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
-        Context ctx = new InitialContext();
-        this.utx = (UserTransaction) ctx.lookup("java:comp/UserTransaction");
-        
         try {
             utx.begin();
             em = getEntityManager();
             Offer persistentOffer = em.find(Offer.class, offer.getNam());
-            Administrato adminIDOld = persistentOffer.getAdminID();
-            Administrato adminIDNew = offer.getAdminID();
-            if (adminIDNew != null) {
-                adminIDNew = em.getReference(adminIDNew.getClass(), adminIDNew.getAdminID());
-                offer.setAdminID(adminIDNew);
+            Collection<Vehicleadvert> vehicleadvertCollectionOld = persistentOffer.getVehicleadvertCollection();
+            Collection<Vehicleadvert> vehicleadvertCollectionNew = offer.getVehicleadvertCollection();
+            Collection<Businessadvert> businessadvertCollectionOld = persistentOffer.getBusinessadvertCollection();
+            Collection<Businessadvert> businessadvertCollectionNew = offer.getBusinessadvertCollection();
+            List<String> illegalOrphanMessages = null;
+            for (Vehicleadvert vehicleadvertCollectionOldVehicleadvert : vehicleadvertCollectionOld) {
+                if (!vehicleadvertCollectionNew.contains(vehicleadvertCollectionOldVehicleadvert)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Vehicleadvert " + vehicleadvertCollectionOldVehicleadvert + " since its offer field is not nullable.");
+                }
             }
+            for (Businessadvert businessadvertCollectionOldBusinessadvert : businessadvertCollectionOld) {
+                if (!businessadvertCollectionNew.contains(businessadvertCollectionOldBusinessadvert)) {
+                    if (illegalOrphanMessages == null) {
+                        illegalOrphanMessages = new ArrayList<String>();
+                    }
+                    illegalOrphanMessages.add("You must retain Businessadvert " + businessadvertCollectionOldBusinessadvert + " since its offer field is not nullable.");
+                }
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
+            }
+            Collection<Vehicleadvert> attachedVehicleadvertCollectionNew = new ArrayList<Vehicleadvert>();
+            for (Vehicleadvert vehicleadvertCollectionNewVehicleadvertToAttach : vehicleadvertCollectionNew) {
+                vehicleadvertCollectionNewVehicleadvertToAttach = em.getReference(vehicleadvertCollectionNewVehicleadvertToAttach.getClass(), vehicleadvertCollectionNewVehicleadvertToAttach.getCode());
+                attachedVehicleadvertCollectionNew.add(vehicleadvertCollectionNewVehicleadvertToAttach);
+            }
+            vehicleadvertCollectionNew = attachedVehicleadvertCollectionNew;
+            offer.setVehicleadvertCollection(vehicleadvertCollectionNew);
+            Collection<Businessadvert> attachedBusinessadvertCollectionNew = new ArrayList<Businessadvert>();
+            for (Businessadvert businessadvertCollectionNewBusinessadvertToAttach : businessadvertCollectionNew) {
+                businessadvertCollectionNewBusinessadvertToAttach = em.getReference(businessadvertCollectionNewBusinessadvertToAttach.getClass(), businessadvertCollectionNewBusinessadvertToAttach.getCode());
+                attachedBusinessadvertCollectionNew.add(businessadvertCollectionNewBusinessadvertToAttach);
+            }
+            businessadvertCollectionNew = attachedBusinessadvertCollectionNew;
+            offer.setBusinessadvertCollection(businessadvertCollectionNew);
             offer = em.merge(offer);
-            if (adminIDOld != null && !adminIDOld.equals(adminIDNew)) {
-                adminIDOld.getOfferCollection().remove(offer);
-                adminIDOld = em.merge(adminIDOld);
+            for (Vehicleadvert vehicleadvertCollectionNewVehicleadvert : vehicleadvertCollectionNew) {
+                if (!vehicleadvertCollectionOld.contains(vehicleadvertCollectionNewVehicleadvert)) {
+                    Offer oldOfferOfVehicleadvertCollectionNewVehicleadvert = vehicleadvertCollectionNewVehicleadvert.getOffer();
+                    vehicleadvertCollectionNewVehicleadvert.setOffer(offer);
+                    vehicleadvertCollectionNewVehicleadvert = em.merge(vehicleadvertCollectionNewVehicleadvert);
+                    if (oldOfferOfVehicleadvertCollectionNewVehicleadvert != null && !oldOfferOfVehicleadvertCollectionNewVehicleadvert.equals(offer)) {
+                        oldOfferOfVehicleadvertCollectionNewVehicleadvert.getVehicleadvertCollection().remove(vehicleadvertCollectionNewVehicleadvert);
+                        oldOfferOfVehicleadvertCollectionNewVehicleadvert = em.merge(oldOfferOfVehicleadvertCollectionNewVehicleadvert);
+                    }
+                }
             }
-            if (adminIDNew != null && !adminIDNew.equals(adminIDOld)) {
-                adminIDNew.getOfferCollection().add(offer);
-                adminIDNew = em.merge(adminIDNew);
+            for (Businessadvert businessadvertCollectionNewBusinessadvert : businessadvertCollectionNew) {
+                if (!businessadvertCollectionOld.contains(businessadvertCollectionNewBusinessadvert)) {
+                    Offer oldOfferOfBusinessadvertCollectionNewBusinessadvert = businessadvertCollectionNewBusinessadvert.getOffer();
+                    businessadvertCollectionNewBusinessadvert.setOffer(offer);
+                    businessadvertCollectionNewBusinessadvert = em.merge(businessadvertCollectionNewBusinessadvert);
+                    if (oldOfferOfBusinessadvertCollectionNewBusinessadvert != null && !oldOfferOfBusinessadvertCollectionNewBusinessadvert.equals(offer)) {
+                        oldOfferOfBusinessadvertCollectionNewBusinessadvert.getBusinessadvertCollection().remove(businessadvertCollectionNewBusinessadvert);
+                        oldOfferOfBusinessadvertCollectionNewBusinessadvert = em.merge(oldOfferOfBusinessadvertCollectionNewBusinessadvert);
+                    }
+                }
             }
             utx.commit();
         } catch (Exception ex) {
@@ -123,11 +189,8 @@ public class OfferJpaController implements Serializable {
         }
     }
 
-    public void destroy(String id) throws NonexistentEntityException, RollbackFailureException, Exception {
+    public void destroy(String id) throws IllegalOrphanException, NonexistentEntityException, RollbackFailureException, Exception {
         EntityManager em = null;
-        Context ctx = new InitialContext();
-        this.utx = (UserTransaction) ctx.lookup("java:comp/UserTransaction");
-        
         try {
             utx.begin();
             em = getEntityManager();
@@ -138,10 +201,23 @@ public class OfferJpaController implements Serializable {
             } catch (EntityNotFoundException enfe) {
                 throw new NonexistentEntityException("The offer with id " + id + " no longer exists.", enfe);
             }
-            Administrato adminID = offer.getAdminID();
-            if (adminID != null) {
-                adminID.getOfferCollection().remove(offer);
-                adminID = em.merge(adminID);
+            List<String> illegalOrphanMessages = null;
+            Collection<Vehicleadvert> vehicleadvertCollectionOrphanCheck = offer.getVehicleadvertCollection();
+            for (Vehicleadvert vehicleadvertCollectionOrphanCheckVehicleadvert : vehicleadvertCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Offer (" + offer + ") cannot be destroyed since the Vehicleadvert " + vehicleadvertCollectionOrphanCheckVehicleadvert + " in its vehicleadvertCollection field has a non-nullable offer field.");
+            }
+            Collection<Businessadvert> businessadvertCollectionOrphanCheck = offer.getBusinessadvertCollection();
+            for (Businessadvert businessadvertCollectionOrphanCheckBusinessadvert : businessadvertCollectionOrphanCheck) {
+                if (illegalOrphanMessages == null) {
+                    illegalOrphanMessages = new ArrayList<String>();
+                }
+                illegalOrphanMessages.add("This Offer (" + offer + ") cannot be destroyed since the Businessadvert " + businessadvertCollectionOrphanCheckBusinessadvert + " in its businessadvertCollection field has a non-nullable offer field.");
+            }
+            if (illegalOrphanMessages != null) {
+                throw new IllegalOrphanException(illegalOrphanMessages);
             }
             em.remove(offer);
             utx.commit();
